Introduce webpack for js bundling
cleanup will follow
