--- djangocms_admin_style/sass/ie.scss
@@ -1,4 +1,0 @@
-#quick-access-closed,
-#quick-access-opened {
-    display: none;
-}
