--- djangocms_admin_style/static/djangocms_admin_style/css/ie.css
@@ -1,1 +1,0 @@
-#quick-access-closed,#quick-access-opened{display:none}
