--- djangocms_admin_style/static/djangocms_admin_style/css/pie.css
@@ -1,11 +1,0 @@
-.pie-element, .bordered, .gradient { behavior: url("/static/cms/PIE.htc"); position: relative; }
-
-.bordered { -moz-border-radius: 5px; -webkit-border-radius: 5px; -o-border-radius: 5px; -ms-border-radius: 5px; -khtml-border-radius: 5px; border-radius: 5px; }
-
-.gradient { background: -webkit-gradient(linear, 50% 0%, 50% 100%, color-stop(0%, #ff0000), color-stop(100%, #0000ff)); background: -webkit-linear-gradient(#ff0000, #0000ff); background: -moz-linear-gradient(#ff0000, #0000ff); background: -o-linear-gradient(#ff0000, #0000ff); -pie-background: linear-gradient(#ff0000, #0000ff); background: linear-gradient(#ff0000, #0000ff); }
-
-.pie-container, .widget { z-index: 0; position: relative; }
-
-.z-pie-element, .widget h3 { behavior: url("/static/cms/PIE.htc"); z-index: 0; }
-
-.has-gradient { behavior: url("/static/cms/PIE.htc"); position: relative; background: -webkit-gradient(linear, 50% 0%, 50% 100%, color-stop(0%, #ff0000), color-stop(100%, #0000ff)); background: -webkit-linear-gradient(#ff0000, #0000ff); background: -moz-linear-gradient(#ff0000, #0000ff); background: -o-linear-gradient(#ff0000, #0000ff); -pie-background: linear-gradient(#ff0000, #0000ff); background: linear-gradient(#ff0000, #0000ff); }
