+++ djangocms_admin_style/templates/admin/base_site.html
@@ -6,8 +6,7 @@
 {% block extrahead %}
 <meta name="viewport" content="width=device-width,initial-scale=1" />
 <link rel="stylesheet" type="text/css" href="{% static 'djangocms_admin_style/css/djangocms-admin.css' %}" />
-<script src="{% static 'djangocms_admin_style/js/datetimefields.js' %}"></script>
-<script src="{% static 'djangocms_admin_style/js/related-widget-wrapper.js' %}"></script>
+<script src="{% static 'djangocms_admin_style/js/dist/bundle.adminstyle.min.js' %}"></script>
 {% include 'admin/inc/extrahead.html' %}{% endblock %}
 
 {% block title %}{% include 'admin/inc/title.html' %}{% endblock %}
