+++ djangocms_admin_style/templates/admin/inc/branding.html
@@ -36,7 +36,3 @@
         {% trans 'Open' %} {% current_site_name %}
     </a>
 </div>
-
-<script src="{% static 'djangocms_admin_style/js/drag-touch-support.js' %}"></script>
-<script src="{% static 'djangocms_admin_style/js/base-admin.js' %}"></script>
-<script src="{% static 'djangocms_admin_style/js/jquery.ui.touch-punch.min.js' %}"></script>
