+++ package.json
@@ -14,9 +14,11 @@
     "gulp-sass": "2.0.3",
     "gulp-sourcemaps": "1.5.2",
     "gulp-util": "3.0.6",
+    "imports-loader": "^0.6.5",
     "minimist": "1.1.1",
     "phantomcss": "^1.0.0",
     "phantomjs-prebuilt": "^2.1.7",
-    "resemblejs": "^2.2.0"
+    "resemblejs": "^2.2.0",
+    "webpack": "^1.13.2"
   }
 }
